- added some translation - added new text styles
