+++ lib/features/persone_screen/register/ui/register_screen.dart
@@ -3,7 +3,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:g_commerce/core/helper/spacer_helper.dart';
 import 'package:g_commerce/features/persone_screen/register/cubit/register_cubit.dart';
-import 'package:g_commerce/features/shared_widgets/form_field/form_field.dart';
+import 'package:g_commerce/features/persone_screen/register/ui/widgets/form_field.dart';
+import 'package:g_commerce/generated/l10n.dart';
 
 class RegisterScreen extends StatelessWidget {
   const RegisterScreen({super.key});
@@ -11,6 +12,9 @@ class RegisterScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.white,
+      ),
       body: SafeArea(
         child: Padding(
             padding: EdgeInsets.symmetric(vertical: 10.h, horizontal: 20.w),
@@ -23,7 +27,7 @@ class RegisterScreen extends StatelessWidget {
                     onPressed: () {
                       validateThenDoSignUp(context);
                     },
-                    child: Text('Register')),
+                    child: Text(S.of(context).register)),
               ],
             )),
       ),
