+++ lib/l10n/intl_ar.arb
@@ -4,6 +4,22 @@
     "food_category" : "أكل",
     "clinics_category" : "عيادات",
     "tech_category" : "تكنولوجيا",
-    "search_bar" : "بحث"
+    "search_bar" : "بحث",
+    "register" : "تسجيل حساب",
+    "login" : "تسجيل الدخول",
+    "user_name" : "اسم المستخدم",
+    "email" : "البريد الالكتروني",
+    "password" : "كلمة السر",
+    "phone" : "رقم الهاتف",
+    "password_validation" : "كلمة السر يجب ان تحتوى على",
+    "uppercase_validation" : "عدد 1 حرف انجليزي كبير",
+    "lowercase_validation" :"عدد 1 حرف انجليزي صغير",
+    "number" : "عدد 1 رقم",
+    "special_character" :"عدد 1 رمز (#,$,%,&)",
+    "at_least_8_character" : "لا تقل طول كلمة السر عن 8",
+    "please_enter_user_name" : "من فضلك ادخل اسم المستخدم",
+    "please_enter_email" : "من فضلك ادخل البريد الالكتروني بشكل صحيح",
+    "please_enter_phone" : "من فضلك ادخل رقم الهاتف بشكل صحيح",
+    "please_enter_password" : "من فضلك ادخل كلمة السر بشكل صحيح"
 
 }
