+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:g_commerce/core/di/dependencey_injection.dart';
 import 'package:g_commerce/core/routing/app_router.dart';
 import 'package:g_commerce/core/routing/routes.dart';
 import 'package:g_commerce/core/theming/app_theme/main_theme.dart';
+import 'package:g_commerce/features/persone_screen/register/ui/register_screen.dart';
 import 'package:g_commerce/features/user_screens/home/screen_ui/home_screen.dart';
 import 'package:g_commerce/generated/l10n.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
